Update Index.zh-Hans.cshtml
Commented out VocaloidOtaku link for reference in case it is to be replaced by something.
Deleted Delicious Cake project link, as it is not available.
japanese -> Japanese
May replace English here with the proofread version after I finish it. Not sure as I'm sure a full Chinese translation is preferred, but I can't do that.