Update Details.cshtml
Don't know if this is how to add the comments tab for Events so I have it all commented out. (It would be all commented out, right?)
That and I'll have to look for the file that has the correct string values, and maybe add them if the string values don't exist yet.